test: split quote and invoice send-email suites
